Update Tailwind configuration, remove unused assets, and enhance UI components
Updated the Tailwind configuration to include a new font family 'ZurichBT' and removed the favicon and OG image files as they are no longer needed. Enhanced the UI components in the Navbar and Footer for better branding, and improved the Hero section with updated content and design elements. Additionally, added a new stakeholders section with a slideshow feature to showcase key partnerships.